requirements.txt
python-dotenv==1.0.1  # For reading environment variables stored in .env file
langchain==0.2.2
langchain-community==0.2.3
langchain-openai==0.1.8  # For embeddings
unstructured==0.14.4  # Document loading
chromadb==0.5.0  # Vector storage
openai==1.31.1  # For embeddings
tiktoken==0.7.0  # For embeddings
nltk==3.8.1
PyPDF2==3.0.1
Flask==3.0.0

# Optional dependencies and platform-specific notes:
# onnxruntime==1.17.1  # Chromadb dependency: on Mac use `conda install onnxruntime -c conda-forge`
# For Windows users, install Microsoft Visual C++ Build Tools first
# Install onnxruntime before installing `chromadb`


# install markdown depenendies with: `pip install "unstructured[md]"` after install the requirements file. Leave this line commented out. 

